refactor extractMetricMetadata into RunUtils
